server/requirements.txt
paho.mqtt
pandas
openpyxl
boto3
dotenv